Fix forcebuild option for PROJ4 modules
The Lib/GIS/Proj/Makefile.PL and Lib/Transform/Proj4/Makefile.PL
would set the $forcebuild option if WITH_PROJ was set in the
perldl.conf file but then would stop the build if auto-detection
failed.  That would give the user no way to work around a problem
in the library detection.  This solution allows the user to
force a build no matter what.  If the options are correct, the
modules will work.